# Conflicts:
#	model/fileManagement/FlySystemManagement.php
#	views/js/controller/actions.js

+++ model/sharedStimulus/css/service/UploadStylesheetService.php
@@ -39,9 +39,22 @@ class UploadStylesheetService extends ConfigurableService
 
         $tmpResource = $uploadedStylesheetDTO->getFileResource();
         $size = filesize($uploadedStylesheetDTO->getTmpFileLink());
-        $this->getStylesheetRepository()->putStream($link, $tmpResource);
-        fclose($tmpResource);
-        unlink($uploadedStylesheetDTO->getTmpFileLink());
+        $this->getStylesheetRepository()->writeStream($link, $tmpResource);
+
+        if (is_resource($tmpResource)) {
+            fclose($tmpResource);
+        } else {
+            $this->logWarning(
+                sprintf(
+                    'Stream for asset stylesheet file "%s" is not valid. It may be already closed',
+                    $uploadedStylesheetDTO->getFileName()
+                )
+            );
+        }
+
+        if (is_writable($uploadedStylesheetDTO->getTmpFileLink())) {
+            unlink($uploadedStylesheetDTO->getTmpFileLink());
+        }
 
         return [
             'alt' => $uploadedStylesheetDTO->getFileName(),
